float en metodoPotencia

+++ src/caraacara.hpp
@@ -12,8 +12,8 @@
 
 // -------------- separador de bajo presupuesto --------------
 // Reduccion de espacio
-int metodoPotencia(const MatrixXf& B, VectorXf& v, int iteraciones);
-void deflacionar(MatrixXf& B, VectorXf& v, int lambda);
+float metodoPotencia(const MatrixXf& B, VectorXf& v, int iteraciones);
+void deflacionar(MatrixXf& B, VectorXf& v, float lambda);
 void transfCaracteristica(MatrixXf& M_x, unsigned int k, unsigned int its, MatrixXf& V);
 int inline norma2(const RowVectorXf& v){ return sqrt(v.dot(v)); };
